backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.7.1
python-dotenv==1.0.1
bcrypt==4.1.3
PyJWT==2.8.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.13.0
nltk==3.8.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
slowapi==0.1.9
python-multipart==0.0.9
openpyxl==3.1.5
greenlet==3.0.3
